Finished passport.js auth work, most logic now in index.js rather than app.js
