+++ package.json
@@ -14,6 +14,7 @@
     "express-session": "^1.15.6",
     "jade": "~1.11.0",
     "morgan": "~1.9.0",
+    "passport": "^0.4.0",
     "react": "^0.14.6",
     "react-dom": "^0.14.6",
     "react-router": "3.0.4",
